+++ src/indicators/mod.rs
@@ -27,3 +27,6 @@ pub use self::average_true_range::AverageTrueRange;
 
 mod moving_average_convergence_divergence;
 pub use self::moving_average_convergence_divergence::MovingAverageConvergenceDivergence;
+
+mod efficiency_ratio;
+pub use self::efficiency_ratio::EfficiencyRatio;