+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+* Implement Kaufman's Efficiency Ratio (ER)
+
 #### v0.1.0 - 2017-12-05
 
 * Initial release